update: image size and css on product detail page
