+++ store/src/index.css
@@ -13,7 +13,7 @@
     --popover: 0 0% 100%;
     --popover-foreground: 222.2 84% 4.9%;
 
-    --primary: 222.2 47.4% 11.2%;
+    --primary: 0 0% 7% /* Original: 222.2 47.4% 11.2% */;
     --primary-foreground: 210 40% 98%;
 
     --secondary: 210 40% 96.1%;
